Updated Unit Tests
